+++ demisto_sdk/commands/common/constants.py
@@ -166,6 +166,14 @@ class FileType(str, Enum):
     TOOL = "tools"
     PACK_METADATA = "packmetadata"
     LAYOUT_RULE = "layoutrule"
+    PIPFILE = "pipfile"
+    TXT = "txt"
+    PIPFILE_LOCK = "pipfilelock"
+    PYLINTRC = "pylintrc"
+    LICENSE = "license"
+    UNIFIED_YML = "unified_yml"
+    INI = "ini"
+    PEM = "pem"
 
 
 RN_HEADER_BY_FILE_TYPE = {