+++ demisto_sdk/tests/constants_test.py
@@ -85,7 +85,7 @@ INVALID_PLAYBOOK_CONDITION_1 = (
 INVALID_IGNORED_UNIFIED_INTEGRATION = (
     f"{GIT_ROOT}/demisto_sdk/tests/test_files/integration_ignored_invalid_unified.yml"
 )
-IGNORED_PNG = f"{GIT_ROOT}/demisto_sdk/tests/test_files/docs_test/closing_params.png"
+IGNORED_PNG = f"{GIT_ROOT}/demisto_sdk/tests/test_files/docs_test/closing-params.png"
 SCRIPT_WITH_PLAYBOOK = "demisto_sdk/tests/test_files/script-with-test-playbook.yml"
 INVALID_PLAYBOOK_CONDITION_2 = (
     f"{GIT_ROOT}/demisto_sdk/tests/test_files/playbook-Invalid_condition_unhandled_"