+++ CHANGELOG.md
@@ -1,7 +1,21 @@
 # Changelog
 ## Unreleased
 * Fixed an issue in **prepare-content** command where large code lines were broken.
+* Fixed an issue where git-*renamed_files* were not retrieved properly.
 
+## 1.9.0
+* Fixed an issue where the Slack notifier was using a deprecated argument.
+* Added the `--docker-image` argument to the **lint** command, which allows determining the docker image to run lint on. Possible options are: `'native:ga'`, `'native:maintenance'`, `'native:dev'`, `'all'`, a specific docker image (from Docker Hub) or, the default `'from-yml'`.
+* Fixed an issue in **prepare-content** command where large code lines were broken.
+* Added a logger warning to **get_demisto_version**, the task will now fail with a more informative message.
+* Fixed an issue where the **upload** and **prepare-content** commands didn't add `fromServerVersion` and `toServerVersion` to layouts.
+* Updated **lint** to use graph instead of id_set when running with `--check-dependent-api-module` flag.
+* Added the marketplaces field to all schemas.
+* Added the flag `--xsoar-only` to the **doc-review** command which enables reviewing documents that belong to XSOAR-supported Packs.
+* Fixed an issue in **update-release-notes** command where an error occurred when executing the same command a second time.
+* Fixed an issue where **validate** would not always ignore errors listed under `.pack-ignore`.
+* Fixed an issue where running **validate** on a specific pack didn't test all the relevant entities.
+* Fixed an issue where fields ending with `_x2` where not replaced in the appropriate Marketplace.
 
 ## 1.8.3
 * Changed **validate** to allow hiding parameters of type 0, 4, 12 and 14 when replacing with type 9 (credentials) with the same name.
@@ -32,7 +46,6 @@
 * Enhanced the **update-content-graph** command to support `--use-git`, `--imported_path` and `--output-path` arguments.
 * Fixed an issue where **doc-review** failed when reviewing command name in some cases.
 * Fixed an issue where **download** didn't identify playbooks properly, and downloaded files with UUIDs instead of file/script names.
-
 
 ## 1.8.1
 * Fixed an issue where **format** created duplicate configuration parameters.